feat: 权限action可在op内操作 #6397
